fixed issue related from c to java

+++ src/utils/filetable.h
@@ -57,5 +57,5 @@ void filetable_destroy(peer_file_table* table);
 void filetable_print(peer_file_table* ptable);
 int file_table_deleteip(peer_file_table* ptable, unsigned int ip);
 int file_table_addip(peer_file_table* ptable,char* name, unsigned int ip);
-unsigned long getmyip();
+unsigned int getmyip();
 void append_node(Node *node, char *body);
